requirements.txt
flask==2.3.3
textblob==0.17.1
transformers==4.33.3
torch==2.0.1
sentencepiece==0.1.99
protobuf==3.20.3
requests==2.31.0
python-dotenv==1.0.0